Try experimental, temporary CircleCI fix
This experimental fix uses an updated version of the build environment to
address the issues arising with Clnag-Format 12. Additionally, as flybys,
the compiler and the tools were upgraded to their most recent versions.
This build environment might therefore not be suitable for getting merged
and may be reverted in future commits.