Temporarily disable work requesting scheduling policy
stdexec's implementation of sync_wait blocks the entire OS thread,
in accordance with P2300 [exec.sync.wait]. Without work stealing,
this causes deadlocks at the PX-thread level, leading to timeouts
in many of the sender-related tests. In the long term, this should
probably be fixed by implementing a custom sync_wait for HPX that
is compatible with stdexec's senders.